Documents/DV-2027-ETH/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dv-portal-ethiopia"
version = "1.0.0"
description = "Django application for DV-2027 program assistance in Ethiopia"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "DV Portal Ethiopia", email = "support@dvportal-ethiopia.com"},
]
keywords = ["django", "diversity-visa", "ethiopia", "telegram-bot"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
]

dependencies = [
    "Django>=4.2.0,<5.0.0",
    "mysqlclient>=2.1.0",
    "django-crispy-forms>=2.0",
    "crispy-bootstrap5>=0.7",
    "pyTelegramBotAPI>=4.14.0",
    "google-generativeai>=0.3.0",
    "Pillow>=10.1.0",
    "requests>=2.31.0",
    "python-dateutil>=2.8.0",
    "python-decouple>=3.8",
    "gunicorn>=21.2.0",
    "structlog>=23.1.0",
    "django-cors-headers>=4.3.0",
    "boto3>=1.28.0",
    "sentry-sdk>=1.32.0",
]

[project.optional-dependencies]
dev = [
    "django-debug-toolbar>=4.2.0",
    "django-extensions>=3.2.0",
]

[project.scripts]
dv-migrate = "migrate_database:main"
dv-collectstatic = "collect_static:main"
dv-createsuperuser = "create_admin:main"
dv-startbot = "start_bot:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["accounts*", "applications*", "core*", "notifications*", "dv_portal*"]

[tool.setuptools.package-data]
"*" = ["*.txt", "*.md", "*.py", "*.html", "*.css", "*.js", "*.png", "*.jpg", "*.svg"]
